+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Fixed
 
 - Fixed an error that could occur when creating a new entry, if the section’s Entry URI Format contained `{canonicalId}`. ([#6080](https://github.com/craftcms/cms/issues/6080))
+- Fixed a bug where `craft\helpers\Db::isTextualColumnType()` was always returning `false`.
 
 ## 3.7.6 - 2021-07-27